Refactor by moving files around dayview.
